docs: document types/workspace and related inconviniences
