+++ frontend/testing-view/src/mocks/chartsConfigurations.ts
@@ -1,4 +1,4 @@
-import type { WorkspaceChartConfig } from "../store/slices/workspacesSlice";
+import type { WorkspaceChartConfig } from "../types/workspace/charts";
 
 export const MOCK_CHARTS: Record<string, WorkspaceChartConfig[]> = {
   "workspace-1": [
