+++ frontend/testing-view/src/types/workspace/sidebar.ts
@@ -1,5 +1,11 @@
 import type { FilterScope } from "./filters";
 
+/**
+ * Sidebar tab identifiers.
+ */
 export type SidebarTab = "commands" | "telemetry";
 
+/**
+ * Record of expanded items per filter scope.
+ */
 export type WorkspaceExpandedItems = Record<FilterScope, Set<number | string>>;
