+++ frontend/testing-view/src/types/workspace/workspace.ts
@@ -1,8 +1,15 @@
 import type { KeyBinding } from "./keyBinding";
 
+/**
+ * Workspace definition.
+ */
 export interface Workspace {
+  /** The name of the workspace */
   name: string;
+  /** The id of the workspace */
   id: string;
+  /** The description of the workspace */
   description: string;
+  /** The key bindings for the workspace */
   keyBindings: KeyBinding[];
 }
